--- src/lib/pusher/index.ts
@@ -1,20 +1,0 @@
-import PusherServer from 'pusher';
-import PusherClient from 'pusher-js';
-
-import { env } from '~/env';
-
-export const pusherServer = new PusherServer({
-  appId: process.env.PUSHER_APP_ID!,
-  key: process.env.PUSHER_APP_KEY!,
-  secret: process.env.PUSHER_APP_SECRET!,
-  cluster: process.env.PUSHER_APP_CLUSTER!,
-  useTLS: true,
-});
-
-export const pusherClient = new PusherClient(env.NEXT_PUBLIC_PUSHER_APP_KEY, {
-  cluster: env.NEXT_PUBLIC_PUSHER_APP_CLUSTER,
-});
-
-export const messageTypes = {
-  NEW_MESSAGE: 'new-message',
-} as const;
