+++ src/lib/pusher/react.ts
@@ -1,0 +1,9 @@
+'use client';
+
+import PusherClient from 'pusher-js';
+
+import { env } from '~/env';
+
+export const pusherClient = new PusherClient(env.NEXT_PUBLIC_PUSHER_APP_KEY, {
+  cluster: env.NEXT_PUBLIC_PUSHER_APP_CLUSTER,
+});
