+++ src/lib/pusher/server.ts
@@ -1,0 +1,11 @@
+import 'server-only';
+
+import PusherServer from 'pusher';
+
+export const pusherServer = new PusherServer({
+  appId: process.env.PUSHER_APP_ID!,
+  key: process.env.PUSHER_APP_KEY!,
+  secret: process.env.PUSHER_APP_SECRET!,
+  cluster: process.env.PUSHER_APP_CLUSTER!,
+  useTLS: true,
+});
