+++ src/lib/pusher/shared.ts
@@ -1,0 +1,3 @@
+export const messageTypes = {
+  NEW_MESSAGE: 'new-message',
+} as const;
